unify all workflow transport tests
